moved build_optimizer to pw._util.py
